third_party/blink/web_tests/external/wpt/html/semantics/forms/constraints/form-validation-validity-valid-expected.txt

This is a testharness.js-based test.
PASS [INPUT in TEXT status] validity.valid must be false if validity.patternMismatch is true
FAIL [INPUT in TEXT status] validity.valid must be false if validity.valueMissing is true assert_true: The validity.valid should be true, when control is disabled. expected true got false
PASS [INPUT in SEARCH status] validity.valid must be false if validity.patternMismatch is true
FAIL [INPUT in SEARCH status] validity.valid must be false if validity.valueMissing is true assert_true: The validity.valid should be true, when control is disabled. expected true got false
PASS [INPUT in TEL status] validity.valid must be false if validity.patternMismatch is true
FAIL [INPUT in TEL status] validity.valid must be false if validity.valueMissing is true assert_true: The validity.valid should be true, when control is disabled. expected true got false
PASS [INPUT in PASSWORD status] validity.valid must be false if validity.patternMismatch is true
FAIL [INPUT in PASSWORD status] validity.valid must be false if validity.valueMissing is true assert_true: The validity.valid should be true, when control is disabled. expected true got false
PASS [INPUT in URL status] validity.valid must be false if validity.patternMismatch is true
PASS [INPUT in URL status] validity.valid must be false if validity.typeMismatch is true
FAIL [INPUT in URL status] validity.valid must be false if validity.valueMissing is true assert_true: The validity.valid should be true, when control is disabled. expected true got false
PASS [INPUT in EMAIL status] validity.valid must be false if validity.patternMismatch is true
PASS [INPUT in EMAIL status] validity.valid must be false if validity.typeMismatch is true
FAIL [INPUT in EMAIL status] validity.valid must be false if validity.valueMissing is true assert_true: The validity.valid should be true, when control is disabled. expected true got false
PASS [INPUT in DATETIME-LOCAL status] validity.valid must be false if validity.rangeOverflow is true
PASS [INPUT in DATETIME-LOCAL status] validity.valid must be false if validity.rangeUnderflow is true
PASS [INPUT in DATETIME-LOCAL status] validity.valid must be false if validity.stepMismatch is true
FAIL [INPUT in DATETIME-LOCAL status] validity.valid must be false if validity.valueMissing is true assert_true: The validity.valid should be true, when control is disabled. expected true got false
PASS [INPUT in DATE status] validity.valid must be false if validity.rangeOverflow is true
PASS [INPUT in DATE status] validity.valid must be false if validity.rangeUnderflow is true
PASS [INPUT in DATE status] validity.valid must be false if validity.stepMismatch is true
FAIL [INPUT in DATE status] validity.valid must be false if validity.valueMissing is true assert_true: The validity.valid should be true, when control is disabled. expected true got false
PASS [INPUT in MONTH status] validity.valid must be false if validity.rangeOverflow is true
PASS [INPUT in MONTH status] validity.valid must be false if validity.rangeUnderflow is true
PASS [INPUT in MONTH status] validity.valid must be false if validity.stepMismatch is true
FAIL [INPUT in MONTH status] validity.valid must be false if validity.valueMissing is true assert_true: The validity.valid should be true, when control is disabled. expected true got false
PASS [INPUT in WEEK status] validity.valid must be false if validity.rangeOverflow is true
PASS [INPUT in WEEK status] validity.valid must be false if validity.rangeUnderflow is true
PASS [INPUT in WEEK status] validity.valid must be false if validity.stepMismatch is true
FAIL [INPUT in WEEK status] validity.valid must be false if validity.valueMissing is true assert_true: The validity.valid should be true, when control is disabled. expected true got false
PASS [INPUT in TIME status] validity.valid must be false if validity.rangeOverflow is true
PASS [INPUT in TIME status] validity.valid must be false if validity.rangeUnderflow is true
PASS [INPUT in TIME status] validity.valid must be false if validity.stepMismatch is true
FAIL [INPUT in TIME status] validity.valid must be false if validity.valueMissing is true assert_true: The validity.valid should be true, when control is disabled. expected true got false
PASS [INPUT in NUMBER status] validity.valid must be false if validity.rangeOverflow is true
PASS [INPUT in NUMBER status] validity.valid must be false if validity.rangeUnderflow is true
PASS [INPUT in NUMBER status] validity.valid must be false if validity.stepMismatch is true
FAIL [INPUT in NUMBER status] validity.valid must be false if validity.valueMissing is true assert_true: The validity.valid should be true, when control is disabled. expected true got false
PASS [INPUT in CHECKBOX status] validity.valid must be false if validity.valueMissing is true
FAIL [INPUT in RADIO status] validity.valid must be false if validity.valueMissing is true assert_false: The validity.valid should be false. expected false got true
PASS [INPUT in FILE status] validity.valid must be false if validity.valueMissing is true
PASS [select]  validity.valid must be false if validity.valueMissing is true
FAIL [textarea]  validity.valid must be false if validity.valueMissing is true assert_true: The validity.valid should be true, when control is disabled. expected true got false
Harness: the test ran to completion.
